Block Traceur MainTvActivity when development options disabled.
The system tracing app is intended to only be accessible when
development options are enabled, but this requirement could be bypassed
by sending an intent to launch the MainTvActivity. This change causes
the activity to immediately exit if development options are disabled.

Bug: 160155846
Test: am start com.android.traceur/com.android.traceur.MainTvActivity
failed when development options were disabled.

Change-Id: I371538eaff180dfc38fa6403a773c24cbbfd4dcf
Merged-In: I371538eaff180dfc38fa6403a773c24cbbfd4dcf